refactor: extract Paycodes, CurrencySettings, RegionalSettings overlays from Dashboard
Extract 3 overlay JSX sections (~549 lines) from Dashboard.js into
dedicated components in components/Settings/:
- PaycodesOverlay.js (~247 lines, includes bech32/LNURL computation)
- CurrencySettingsOverlay.js (~202 lines, includes currency filtering)
- RegionalSettingsOverlay.js (~239 lines, number/bitcoin/numpad formats)

Clean up Dashboard.js imports: remove QRCode, bech32, and 10
number-format exports now handled by extracted components.

Dashboard.js: 7,998 -> 7,449 lines. Build passes.
